+++ src/Schema/resolvers/mutation/deleteMouse.ts
@@ -1,6 +1,11 @@
 import MouseModel from "../../../models/mouse";
+import { AuthenticationError } from 'apollo-server-express';
 
-export const deleteMouse = async (_: void, args: { id: string }): Promise<boolean> => {
+export const deleteMouse = async (_: void, args: { id: string }, context: any): Promise<boolean> => {
+    if (!context.user) {
+        throw new AuthenticationError('Usuario no autenticado');
+    }
+
     const { id } = args;
     
     const deletedMouse = await MouseModel.findByIdAndDelete(id);
